Анимации

+++ src/stages/rooms/abstract-room.stage.js
@@ -1,10 +1,10 @@
-import {Game, Sprite} from 'jetcode-scrubjs';
-import {ButtonSprite} from "../../sprites/button.sprite";
-import {MonitorStage} from "../monitor.stage";
-import {AbstractStage} from "../abstract.stage";
-import {GameState} from "../../services/game.state";
-import {BlankButtonSprite} from "../../sprites/blank-button.sprite";
-import {LongButtonSprite} from "../../sprites/long-button.sprite";
+import { Game, Sprite } from 'jetcode-scrubjs';
+import { ButtonSprite } from "../../sprites/button.sprite";
+import { MonitorStage } from "../monitor.stage";
+import { AbstractStage } from "../abstract.stage";
+import { GameState } from "../../services/game.state";
+import { BlankButtonSprite } from "../../sprites/blank-button.sprite";
+import { LongButtonSprite } from "../../sprites/long-button.sprite";
 
 export class AbstractRootStage extends AbstractStage {
 	thumbnail;
@@ -20,15 +20,17 @@ export class AbstractRootStage extends AbstractStage {
 	robotChargeDif = 0;
 
 	isCleaningByRobot = false;
-
+	visualizers = [];
+	visualiserMoving = true;
 	isVisualized = false;
 
+	visualizerType;
 
 	init() {
 		super.init()
 		this.gameState = GameState.getInstance();
 
-		this.onReady(()=>{
+		this.onReady(() => {
 			this.monitorStage = MonitorStage.getInstance();
 		})
 
@@ -54,29 +56,30 @@ export class AbstractRootStage extends AbstractStage {
 		this.forever(this.gameTickAllRooms, 1000);
 		this.forever(this.updateHelpText);
 
-		this.visualiser = new Sprite();
-		this.visualiser.hidden = true;
-		this.visualiser.moving = true;
+		// this.visualiser = new Sprite();
+		// this.setVisCostumes();
+		// this.visualiser.hidden = true;
+		// this.visualiser.moving = true;
 
 		this.onStart(() => {
 			this.visualizerSpawn();
 
-            if (this.gameState.visits[this.getRoomType()] && this.gameState.teachingMode) {
+			if (this.gameState.visits[this.getRoomType()] && this.gameState.teachingMode) {
 				this.gameState.visits[this.getRoomType()] = false;
 
 				this.helper.show(this.getHelpText(), GameState.BOSS_PERSON);
 				this.helper.setButtonText('Дальше');
-				this.helper.onClick(()=>{
+				this.helper.onClick(() => {
 					this.showInstructionDialog();
 				})
-            }
+			}
 		});
 
 		this.onReady(() => {
 			this.resetRoom();
 		});
 
-		this.forever(()=>{
+		this.forever(() => {
 			if (this.currentQuantity <= 0) {
 				this.isVisualized = false;
 			}
@@ -88,7 +91,7 @@ export class AbstractRootStage extends AbstractStage {
 		getManureButton.x = 430;
 		getManureButton.y = 540;
 
-		getManureButton.onReady(()=>{
+		getManureButton.onReady(() => {
 			getManureButton.setLabel('Собрать помет');
 		});
 
@@ -175,14 +178,14 @@ export class AbstractRootStage extends AbstractStage {
 	}
 
 	resetRoom() {
-		this.currentQuantity = 0;
+		this.currentQuantity = 1;
 		this.isRoomReady = false;
 		this.pollution = 0;
 	}
 
-	roomTick () {
+	roomTick() {
 		if (this.robot) {
-			if (this.pollution > this.cleaningValue){
+			if (this.pollution > this.cleaningValue) {
 				this.pollution -= this.cleaningValue;
 				this.gameState.manure += this.cleaningValue;
 			} else {
@@ -204,7 +207,7 @@ export class AbstractRootStage extends AbstractStage {
 		this.thumbnail = thumbnailSprite;
 	}
 
-	gameTickAllRooms () {
+	gameTickAllRooms() {
 		MonitorStage.getInstance().gameTick();
 	}
 
@@ -216,81 +219,100 @@ export class AbstractRootStage extends AbstractStage {
 			return;
 		}
 
-		this.visualiser.deleteClones();
+		if (!this.visualizerType) {
+			return
+		}
+
+		for (const vis of this.visualizers) {
+			vis.delete();
+		}
 
 		for (let i = 0; i < this.currentQuantity; i++) {
-			let visClone = this.visualiser.createClone();
+			const visClone = new Sprite(this);
 
+			visClone.typeNum = this.game.getRandom(0, this.gameState.objectTypesCount[this.visualizerType] - 1);
+			visClone.activities = this.gameState.getAnimationsNames(this.visualizerType);
+			for(const activity of visClone.activities) {
+				for(const costume of this.gameState.getAnimationCostumes(this.visualizerType, activity, visClone.typeNum))
+				{
+					visClone.addCostume(costume)
+				}
+			}
 
-			visClone.hidden = false;
-			visClone.x = this.game.getRandom(200, 400)
-			visClone.y = this.game.getRandom(200, 350)
-			visClone.layer = 4
-			visClone.rotateStyle = 'leftRight'
-			visClone.size = 150
-			visClone.xSpeed = this.game.getRandom(-10, 10) / 10
-			visClone.ySpeed = this.game.getRandom(-10, 10) / 10
+			visClone.currentActivity = 0;	
 			visClone.number = i + 1;
 
-			let randCos = this.game.getRandom(0, this.visualiser.costumes.length)
-			visClone.switchCostume(randCos)
+			this.prepareVis(visClone);
 
+			// let randCos = this.game.getRandom(0, this.visualiser.costumes.length)
+			// visClone.switchCostume(randCos)
 
-			if (this.visualiser.moving) {
-
-				visClone.forever(function () {
+			visClone.forever(() => {
+				if (this.gameState.objectAnimationInfo[this.visualizerType][visClone.activities[visClone.currentActivity]].moving){
 
 					visClone.x += visClone.xSpeed
 					visClone.y += visClone.ySpeed
-
+					
 					if (visClone.xSpeed > 0) {
 						visClone.direction = -90;
 					} else {
 						visClone.direction = 90;
 					}
-
+					
 					if (visClone.x > 450) {
 						visClone.xSpeed *= -1
 					}
-
+					
 					if (visClone.x < 150) {
 						visClone.xSpeed *= -1
 					}
-
+					
 					if (visClone.y > 380) {
 						visClone.ySpeed *= -1
 					}
-
+					
 					if (visClone.y < 200) {
 						visClone.ySpeed *= -1
 					}
-
-				})
-			}
-			visClone.forever(()=>{
+				}
 				if (visClone.number > this.currentQuantity) {
 					visClone.delete();
 				}
 			})
 
-		}
+			visClone.forever(()=>{
+				visClone.nextCostume(this.gameState.objectAnimationInfo[this.visualizerType][visClone.activities[visClone.currentActivity]].min, this.gameState.objectAnimationInfo[this.visualizerType][visClone.activities[visClone.currentActivity]].max);
+				if (visClone.getCostumeIndex() == this.gameState.objectAnimationInfo[this.visualizerType][visClone.activities[visClone.currentActivity]].max) {
+					visClone.currentActivity = this.game.getRandom(0, visClone.activities.length - 1)
+				}
+			}, 100)
+
+			visClone.onReady(()=>{
+				visClone.run(); //Да, костыль, но не работает по другому
+			})
+			this.visualizers.push(visClone)
+		}	
 
 		this.isVisualized = true;
 	}
 
 	//Перезаписать в классе комнаты, чтобы задать нужные костюмы в спрайт Visualiser
 	setVisCostumes() {
-		
+
+	}
+
+	getVisType() {
+		console.error('Визуализатор не определен.');
 	}
 
 	showInstructionDialog() {
 		this.helper.show(this.getInstructionText(), GameState.BOSS_PERSON);
 		this.helper.setButtonText('Дальше');
 
-		this.helper.onClick(()=>{
+		this.helper.onClick(() => {
 			this.helper.show(GameState.getInstance().getHeroAnswer(GameState.NORMAL_PERSON_EMOTION) + ', босс!', GameState.HERO_PERSON);
 			this.helper.setButtonText('Конец');
-			this.helper.onClick(()=>{this.helper.hide()});
+			this.helper.onClick(() => { this.helper.hide() });
 		});
 	}
 
@@ -313,4 +335,15 @@ export class AbstractRootStage extends AbstractStage {
 
 		this.showHelp(helpText);
 	}
+
+	prepareVis(sprite){
+        sprite.hidden = false;
+        sprite.x = sprite.game.getRandom(200, 400)
+        sprite.y = sprite.game.getRandom(200, 350)
+        sprite.layer = 4
+        sprite.rotateStyle = 'leftRight'
+        sprite.size = 150
+        sprite.xSpeed = sprite.game.getRandom(-10, 10) / 10
+        sprite.ySpeed = sprite.game.getRandom(-10, 10) / 10
+    }
 }
